Getting data from cloud storage/requirements.txt
# Function dependencies, for example:
# package>=version
pandas==1.1.3
google-cloud-bigquery==1.24.0
google-cloud-storage==1.26.0
pandas==1.1.3
gcsfs==0.7.1
pyarrow==1.0.0
numpy==1.19.2
requests==2.20.0
requests-toolbelt==0.9.1